move modules into assets
rm import
fix improper li nesting
break counts apparently
